GridTrack optimise

+++ avr/cores/megacommand/MCL/GridTrack.h
@@ -70,7 +70,7 @@ public:
   ///  caller guarantees that the type is reconstructed correctly
   ///  downloads from BANK1 to the runtime object
   bool load_from_mem(uint8_t column, size_t size = 0) {
-    uint32_t bytes = size ? size : get_track_size();
+    uint16_t bytes = size ? size : get_track_size();
     uint32_t pos = get_region() + get_region_size() * (uint32_t)(column);
     volatile uint8_t *ptr = reinterpret_cast<uint8_t *>(pos);
     memcpy_bank1(this, ptr, bytes);
